+++ src/main/java/com/example/shop_online/entity/GoodsDetail.java
@@ -45,7 +45,7 @@ public class GoodsDetail {
     @ApiModelProperty("逻辑删除(0-未删除，1-已删除)")
     @TableField("delete_flag")
     @TableLogic
-    private Integer deleteFlag;
+    private Byte deleteFlag;
 
     @ApiModelProperty("创建时间")
     @TableField(value = "create_time", fill = FieldFill.INSERT)
